Reduce parameters for encoding limits tests
