+++ truffle/src/com.oracle.truffle.api.bytecode.test/src/com/oracle/truffle/api/bytecode/test/basic_interpreter/BasicInterpreterTest.java
@@ -65,7 +65,6 @@ import org.junit.runners.Parameterized;
 
 import com.oracle.truffle.api.RootCallTarget;
 import com.oracle.truffle.api.bytecode.BytecodeConfig;
-import com.oracle.truffle.api.bytecode.BytecodeEncodingException;
 import com.oracle.truffle.api.bytecode.BytecodeLabel;
 import com.oracle.truffle.api.bytecode.BytecodeLocal;
 import com.oracle.truffle.api.bytecode.BytecodeNode;
@@ -1948,6 +1947,7 @@ public class BasicInterpreterTest extends AbstractBasicInterpreterTest {
         assertEquals(1337L, root.call());
     }
 
+    @Test
     public void testVariadicFallback() {
         // return variadicOperation(arg0, arg1, arg2);
 
@@ -2236,290 +2236,6 @@ public class BasicInterpreterTest extends AbstractBasicInterpreterTest {
                 b.endRoot();
             });
         });
-    }
-
-    @Test
-    public void testManyBytecodes() {
-        BasicInterpreter node = parseNode("manyBytecodes", b -> {
-            b.beginRoot();
-            b.beginBlock();
-            for (int i = 0; i < Short.MAX_VALUE * 2; i++) {
-                b.emitLoadConstant(123L);
-            }
-            b.beginReturn();
-            b.emitLoadConstant(42L);
-            b.endReturn();
-            b.endBlock();
-            b.endRoot();
-        });
-
-        assertEquals(42L, node.getCallTarget().call());
-    }
-
-    @Test
-    public void testManyConstants() {
-        BasicInterpreter node = parseNode("manyConstants", b -> {
-            b.beginRoot();
-            b.beginBlock();
-            for (int i = 0; i < Short.MAX_VALUE * 2; i++) {
-                b.emitLoadConstant((long) i);
-            }
-            b.beginReturn();
-            b.emitLoadConstant(42L);
-            b.endReturn();
-            b.endBlock();
-            b.endRoot();
-        });
-
-        assertEquals(42L, node.getCallTarget().call());
-    }
-
-    @Test
-    public void testManyNodes() {
-        BasicInterpreter node = parseNode("manyNodes", b -> {
-            b.beginRoot();
-            b.beginBlock();
-            for (int i = 0; i < Short.MAX_VALUE * 2; i++) {
-                b.emitVoidOperation();
-            }
-            b.beginReturn();
-            b.emitLoadConstant(42L);
-            b.endReturn();
-            b.endBlock();
-            b.endRoot();
-        });
-
-        assertEquals(42L, node.getCallTarget().call());
-    }
-
-    @Test
-    public void testManyConditionalBranches() {
-        BasicInterpreter node = parseNode("manyConditionalBranches", b -> {
-            b.beginRoot();
-            b.beginBlock();
-            for (int i = 0; i < Short.MAX_VALUE * 2; i++) {
-                b.beginConditional();
-                b.emitLoadArgument(0);
-                b.emitLoadConstant(123L);
-                b.emitLoadConstant(321L);
-                b.endConditional();
-            }
-            b.beginReturn();
-            b.emitLoadConstant(42L);
-            b.endReturn();
-            b.endBlock();
-            b.endRoot();
-        });
-
-        assertEquals(42L, node.getCallTarget().call(true));
-    }
-
-    @Test
-    public void testManyLocals() {
-        BasicInterpreter node = parseNode("manyLocals", b -> {
-            b.beginRoot();
-            b.beginBlock();
-
-            for (int i = 0; i < Short.MAX_VALUE - 10; i++) {
-                b.createLocal();
-            }
-            BytecodeLocal x = b.createLocal();
-            b.beginStoreLocal(x);
-            b.emitLoadConstant(42L);
-            b.endStoreLocal();
-
-            b.beginReturn();
-            b.emitLoadLocal(x);
-            b.endReturn();
-            b.endBlock();
-            b.endRoot();
-        });
-        assertEquals(42L, node.getCallTarget().call());
-    }
-
-    @Test
-    public void testTooManyLocals() {
-        assertThrows(BytecodeEncodingException.class, () -> {
-            parseNode("tooManyLocals", b -> {
-                b.beginRoot();
-                b.beginBlock();
-
-                for (int i = 0; i <= 0xffff; i++) {
-                    b.createLocal();
-                }
-                BytecodeLocal x = b.createLocal();
-                b.beginStoreLocal(x);
-                b.emitLoadConstant(42L);
-                b.endStoreLocal();
-
-                b.beginReturn();
-                b.emitLoadLocal(x);
-                b.endReturn();
-                b.endBlock();
-                b.endRoot();
-            });
-        });
-    }
-
-    @Test
-    public void testLargeStackPointer() {
-        BasicInterpreter node = parseNode("largeStackPointer", b -> {
-            b.beginRoot();
-            b.beginBlock();
-
-            BytecodeLocal x = b.createLocal();
-            int localCount = 0x10000 - run.getFrameBaseSlots();
-            for (int i = 1; i < localCount; i++) {
-                b.createLocal();
-            }
-            b.beginStoreLocal(x);
-            b.emitLoadConstant(42L);
-            b.endStoreLocal();
-
-            // The operand stack begins above the maximum unsigned-short frame index. This
-            // expression must not overwrite the low-index local.
-            b.beginVariadicOperation();
-            for (int i = 0; i < 4; i++) {
-                b.emitLoadConstant(0L);
-            }
-            b.endVariadicOperation();
-
-            b.beginReturn();
-            b.emitLoadLocal(x);
-            b.endReturn();
-            b.endBlock();
-            b.endRoot();
-        });
-        assertEquals(42L, node.getCallTarget().call());
-    }
-
-    @Test
-    public void testManyRoots() {
-        BytecodeRootNodes<BasicInterpreter> nodes = createNodes(BytecodeConfig.DEFAULT, b -> {
-            for (long i = 0; i <= Short.MAX_VALUE + 1; i++) {
-                b.beginRoot();
-                BytecodeLocal local = b.createLocal();
-                b.beginStoreLocal(local);
-                b.emitLoadConstant(i);
-                b.endStoreLocal();
-                b.beginReturn();
-                // Materialized accesses use root index. Ensure high root indices can be encoded.
-                b.beginLoadLocalMaterialized(local);
-                b.emitMaterializeFrame();
-                b.endLoadLocalMaterialized();
-                b.endReturn();
-                b.endRoot();
-            }
-        });
-        assertEquals(0L, nodes.getNode(0).getCallTarget().call());
-        assertEquals(42L, nodes.getNode(42).getCallTarget().call());
-        assertEquals((long) (Short.MAX_VALUE - 1), nodes.getNode(Short.MAX_VALUE - 1).getCallTarget().call());
-        assertEquals((long) Short.MAX_VALUE + 1, nodes.getNode(Short.MAX_VALUE + 1).getCallTarget().call());
-
-        BytecodeRootNodes<BasicInterpreter> roundTripped = doRoundTrip(nodes);
-        assertEquals((long) Short.MAX_VALUE + 1, roundTripped.getNode(Short.MAX_VALUE + 1).getCallTarget().call());
-    }
-
-    @Test
-    public void testManyInstructionsInLoop() {
-        BasicInterpreter node = parseNode("manyInstructionsInLoop", b -> {
-            b.beginRoot();
-            b.beginBlock();
-
-            BytecodeLocal x = b.createLocal();
-            b.beginStoreLocal(x);
-            b.emitLoadConstant(0L);
-            b.endStoreLocal();
-
-            BytecodeLocal result = b.createLocal();
-
-            b.beginStoreLocal(result);
-            b.emitLoadConstant(0L);
-            b.endStoreLocal();
-
-            b.beginWhile();
-            b.beginLess();
-            b.emitLoadLocal(x);
-            b.emitLoadConstant(5L);
-            b.endLess();
-
-            b.beginBlock();
-            for (int i = 0; i < Short.MAX_VALUE * 2; i++) {
-                b.emitVoidOperation();
-            }
-            // x = x + 1
-            b.beginStoreLocal(x);
-            b.beginAdd();
-            b.emitLoadLocal(x);
-            b.emitLoadConstant(1L);
-            b.endAdd();
-            b.endStoreLocal();
-
-            // result += x
-            b.beginStoreLocal(result);
-            b.beginAdd();
-            b.emitLoadLocal(result);
-            b.emitLoadLocal(x);
-            b.endAdd();
-            b.endStoreLocal();
-
-            b.endBlock();
-
-            b.endWhile();
-
-            b.beginReturn();
-            b.emitLoadLocal(result);
-            b.endReturn();
-            b.endBlock();
-            b.endRoot();
-        });
-
-        assertEquals(15L, node.getCallTarget().call());
-    }
-
-    @Test
-    public void testManyStackValues() {
-        BasicInterpreter node = parseNode("manyStackValues", b -> {
-            b.beginRoot();
-            b.beginReturn();
-            for (int i = 0; i < Short.MAX_VALUE - 1; i++) {
-                b.beginAdd();
-                b.emitLoadConstant(1L);
-            }
-            b.emitLoadConstant(0L);
-
-            for (int i = 0; i < Short.MAX_VALUE - 1; i++) {
-                b.endAdd();
-            }
-
-            b.endReturn();
-            b.endRoot();
-        });
-
-        assertEquals((long) Short.MAX_VALUE - 1, node.getCallTarget().call());
-    }
-
-    @Test
-    public void testTooManyStackValues() {
-        assertThrowsWithMessage("Maximum stack height exceeded", BytecodeEncodingException.class, () -> {
-            parseNode("tooManyStackValues", b -> {
-                b.beginRoot();
-                b.beginReturn();
-                for (int i = 0; i < Short.MAX_VALUE; i++) {
-                    b.beginAdd();
-                    b.emitLoadConstant(1L);
-                }
-                b.emitLoadConstant(0L);
-
-                for (int i = 0; i < Short.MAX_VALUE; i++) {
-                    b.endAdd();
-                }
-
-                b.endReturn();
-                b.endRoot();
-            });
-        });
-
     }
 
     @Test
